Move client test packages
