+++ src/test/java/pt/ua/deti/tqs/backend/functional/client/ClientCucumberTest.java
@@ -1,4 +1,4 @@
-package pt.ua.deti.tqs.backend.functional;
+package pt.ua.deti.tqs.backend.functional.client;
 
 import org.junit.platform.suite.api.ConfigurationParameter;
 import org.junit.platform.suite.api.IncludeEngines;
@@ -10,7 +10,7 @@ import static io.cucumber.junit.platform.engine.Constants.GLUE_PROPERTY_NAME;
 @SuppressWarnings("java:S2187")
 @Suite
 @IncludeEngines("cucumber")
-@SelectClasspathResource("pt/ua/deti/tqs/backend/functional")
-@ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "pt.ua.deti.tqs.backend.functional")
-public class CucumberTest {
+@SelectClasspathResource("pt/ua/deti/tqs/backend/functional/client")
+@ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "pt.ua.deti.tqs.backend.functional.client")
+public class ClientCucumberTest {
 }
