+++ src/test/java/pt/ua/deti/tqs/backend/functional/client/LoginSteps.java
@@ -22,7 +22,7 @@ public class LoginSteps {
     @Given("I navigate to {string}")
     public void iNavigateTo(String url) {
         FirefoxOptions options = new FirefoxOptions();
-        // options.addArguments("-headless");
+        options.addArguments("-headless");
 
         driver = new FirefoxDriver(options);
         driver.get(url);
